Restore Analog Mic Boost on quit on XO-{1.75,4}
Restore the Analog Mic Boost control to the default when the activity
quits.

Audio recording level on XO-1.75 and XO-4 is set by two ALSA mixer
controls, Capture and Analog Mic Boost.  Capture is exposed on the Sugar
Frame as Microphone.  Analog Mic Boost is not exposed, was set to zero
by Measure activity quit (because on earlier models it was a switch, not
a level), and persists at zero over reboots thanks to alsactl.

Test case; try to use Record to capture voice after quit of Measure.

http://dev.laptop.org/ticket/12867

Signed-off-by: James Cameron <quozl@laptop.org>